pyproject.toml
[project]
name = "ruwiki-rag"
version = "0.1.0"
description = "Russian wikipedia dump RAG system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.10.1",
    "einops>=0.8.1",
    "jupyter>=1.1.1",
    "langchain-community>=0.3.29",
    "langchain-gigachat>=0.3.12",
    "langchain-huggingface>=0.3.1",
    "langchain-qdrant>=0.2.1",
    "lxml>=6.0.2",
    "qdrant-client>=1.15.1",
    "sentence-transformers==5.1.1",
    "torch>=2.8.0",
    "transformers==4.51.0",
    "wikiextractor>=3.0.6",
]

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu129", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[tool.uv.extra-build-dependencies]
flash-attn = ["torch"]
